backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
paho-mqtt
pydantic
sqlalchemy
